Improve dynamic resizing of number in display.
If there is only one digit (as after a board soft or hard reset) set the font size to default.
Now, instead of firing watches to bring the digitBox down it the board limits, when the digitBox width exceeds the maxWidth, a loop is run to quickly bring it down within the same watch detection! Totally avoiding the issue of running into the 10 digest limit! 🤓