Integrate the percentage calculation
The operation can be: 
(1) Unary: there is one number in the stack. Press %. The result is the percentage version of that number. Result pushed into number stack.
(2) Binary: there are two number on the stack. There is an ongoing operation. Press %.  The second number is used as the percentage to calculate from the first number. The ongoing operation is then calculated between the first number and the calculated percentage. For example: 4 + 5 % yields 4.2 (4 + (5% of 4)).