Add logic to flip sign of current active number or last input number
If the number stack is empty, the number on the board is the current active number, change sign freely.
If a number has been just entered and no new number has been started, the sign can still be flipped: this could be the result of an operation or a number just entered. 
If there is a number already on the stack, and a new number is being entered, the sign of the active number can be flipped.