Create a workflow to give priority to * and / over + and -
A * or / immediately requested after an ongoing + or -, will place the + or - in stand by and perform the * or / or a chain of * and / till = or another + or - is requested. Once the */ chain is broken, the */ operations are computed and then the standby + or - is performed with the resultant of the */ chain. 🤓